requirements.txt
numpy==2.3.4
pygame==2.6.1
PyOpenGL==3.1.10